Big spike to make submitting HALO forms work
Spike a first version of a hypermedia client for CLJS. Also spike an
initial version of yaks-sinatra, this needs to become its own gem.

There's a bunch of accidental complexity here to make it work, in
particular cross domain, I had to set CORS headers, make sure we do
non-authenticated XHR requests, and rewrite URLs in the response to be
absolute. This last bit could nicely become some Yaks middleware. This
bit also depends on a change that I still need to push out in Yaks where
hooks receive the rack env.